# swagger/swagger.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:9090
...
paths:
...
  /api/v1/skill:
    post:
      operationId: createSkill
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillCreateDto"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SkillDto"
          description: Skill created
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Create a new skill
      tags:
      - Skill
    put:
      operationId: updateSkill
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpdateDto"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SkillDto"
          description: Skill updated
        "400":
          description: Invalid input or skill not found
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Update an existing skill
      tags:
      - Skill
  /api/v1/skill/{id}:
    delete:
      operationId: deleteSkill
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Skill deleted
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          description: Skill not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Delete a skill by ID
      tags:
      - Skill
    get:
      operationId: getSkillById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SkillDto"
          description: Skill found
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Get a skill by ID
      tags:
      - Skill
...
components:
  schemas:
    BadRequestErrorResponse:
      type: object
      description: Bad Request Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "400"
        message:
          type: string
          description: Description of the error
          example: Invalid input provided
...
    ConflictErrorResponse:
      type: object
      description: Conflict Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "409"
        message:
          type: string
          description: Description of the error
          example: Data Conflict
...
    InternalServerErrorResponse:
      type: object
      description: Internal Server Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "500"
        message:
          type: string
          description: Description of the error
          example: Internal Server Error
...
    NotFoundErrorResponse:
      type: object
      description: Not Found Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "404"
        message:
          type: string
          description: Description of the error
          example: Not Found
...
    SkillCreateDto:
      type: object
      description: Skill Create Model
      properties:
        name:
          type: string
          description: Name of the skill
        proficiency:
          type: string
          description: Proficiency level of the skill
          enum:
          - BEGINNER
          - INTERMEDIATE
          - ADVANCED
          - EXPERT
        resumeId:
          type: integer
          format: int64
          description: ID of the resume the skill is associated with
      required:
      - name
    SkillDto:
      type: object
      description: Skill Model
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the skill
        name:
          type: string
          description: Name of the skill
        proficiency:
          type: string
          description: Proficiency level of the skill
          enum:
          - BEGINNER
          - INTERMEDIATE
          - ADVANCED
          - EXPERT
    SkillUpdateDto:
      type: object
      description: Skill Update Model
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the skill
        name:
          type: string
          description: Name of the skill
        proficiency:
          type: string
          description: Proficiency level of the skill
          enum:
          - BEGINNER
          - INTERMEDIATE
          - ADVANCED
          - EXPERT
        resumeId:
          type: integer
          format: int64
          description: Resume ID associated with the skill
...